requirements.txt
fastapi
uvicorn
anthropic
curl-cffi
python-dotenv
rich
